+++ source/features/latest-tag-button.tsx
@@ -9,8 +9,9 @@ import * as pageDetect from 'github-url-detection';
 import * as api from '../libs/api';
 import features from '../libs/features';
 import fetchDom from '../libs/fetch-dom';
+import parseRoute from '../github-helpers/parse-route';
 import getDefaultBranch from '../libs/get-default-branch';
-import {getRepoURL, getCurrentBranch, getRepoGQL, getLatestVersionTag, parseRoute} from '../libs/utils';
+import {getRepoURL, getCurrentBranch, getRepoGQL, getLatestVersionTag} from '../libs/utils';
 
 interface RepoPublishState {
 	latestTag: string | false;
@@ -95,9 +96,9 @@ async function init(): Promise<false | void> {
 	if (pageDetect.isRepoRoot()) {
 		href = `/${getRepoURL()}/tree/${latestTag}`;
 	} else {
-		const pathnameParts = parseRoute(location.pathname);
-		pathnameParts.branch = latestTag;
-		href = pathnameParts.toString();
+		const path = parseRoute(location.pathname);
+		path.branch = latestTag;
+		href = path.toString();
 	}
 
 	const link = (
