+++ source/libs/utils.ts
@@ -156,35 +156,6 @@ export function getLatestVersionTag(tags: string[]): string {
 	return latestVersion;
 }
 
-interface Pathname {
-	user: string;
-	repository: string;
-	route: string;
-	branch: string;
-	filePath: string;
-	toString: () => string;
-}
-export function parseRoute(pathname: string): Pathname {
-	const [user, repository, route, ...next] = pathname.replace(/^\/|\/$/g, '').split('/');
-	const parts = next.join('/');
-	const branch = getCurrentBranch();
-	if (parts !== branch && !parts.startsWith(branch + '/')) {
-		throw new Error('The branch of the current page must match the branch in the `pathname` parameter');
-	}
-
-	const filePath = parts.replace(branch, '').replace(/^\//, '');
-	return {
-		user,
-		repository,
-		route,
-		branch,
-		filePath,
-		toString() {
-			return `/${this.user}/${this.repository}/${this.route}/${this.branch}/${this.filePath}`.replace(/\/$/, '');
-		}
-	};
-}
-
 const escapeRegex = (string: string) => string.replace(/[\\^$.*+?()[\]{}|]/g, '\\$&');
 export const prCommitRegex = new RegExp(`\\b${escapeRegex(location.origin)}[/][^/]+[/][^/]+[/]pull[/]\\d+[/]commits[/][0-9a-f]{7,40}\\b(?! \\]|\\))`, 'gi');
